rename file to avoid collision
